// packages/sdk/src/generated/types/terms.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getAddressDecoder,
  getAddressEncoder,
  getI64Decoder,
  getI64Encoder,
  getOptionDecoder,
  getOptionEncoder,
  getStructDecoder,
  getStructEncoder,
  getU64Decoder,
  getU64Encoder,
  getU8Decoder,
  getU8Encoder,
  transformEncoder,
  type Address,
  type Codec,
  type Decoder,
  type Encoder,
  type Option,
  type OptionOrNullable,
} from "@solana/kit";
import {
  getLimitDecoder,
  getLimitEncoder,
  getReceiveDecoder,
  getReceiveEncoder,
  type Limit,
  type LimitArgs,
  type Receive,
  type ReceiveArgs,
} from "./index.ts";

/** What an authority permits: the same for a policy and for a signed intent. */
export type Terms = {
  version: number;
  /** 0 mainnet, 1 devnet, 2 testnet, 3 localnet: the cluster the program was built for. */
  cluster: number;
  authority: Address;
  /** Who may pull. None is anyone. */
  spender: Option<Address>;
  notBefore: bigint;
  notAfter: Option<bigint>;
  /** A signed intent's nonce. A policy does not use it. */
  salt: bigint;
  limit: Limit;
  receive: Option<Receive>;
};

export type TermsArgs = {
  /** 0 mainnet, 1 devnet, 2 testnet, 3 localnet: the cluster the program was built for. */
  cluster: number;
  authority: Address;
  /** Who may pull. None is anyone. */
  spender: OptionOrNullable<Address>;
  notBefore: number | bigint;
  notAfter: OptionOrNullable<number | bigint>;
  /** A signed intent's nonce. A policy does not use it. */
  salt: number | bigint;
  limit: LimitArgs;
  receive: OptionOrNullable<ReceiveArgs>;
};

export function getTermsEncoder(): Encoder<TermsArgs> {
  return transformEncoder(
    getStructEncoder([
      ["version", getU8Encoder()],
      ["cluster", getU8Encoder()],
      ["authority", getAddressEncoder()],
      ["spender", getOptionEncoder(getAddressEncoder())],
      ["notBefore", getI64Encoder()],
      ["notAfter", getOptionEncoder(getI64Encoder())],
      ["salt", getU64Encoder()],
      ["limit", getLimitEncoder()],
      ["receive", getOptionEncoder(getReceiveEncoder())],
    ]),
    (value) => ({ ...value, version: 1 }),
  );
}

export function getTermsDecoder(): Decoder<Terms> {
  return getStructDecoder([
    ["version", getU8Decoder()],
    ["cluster", getU8Decoder()],
    ["authority", getAddressDecoder()],
    ["spender", getOptionDecoder(getAddressDecoder())],
    ["notBefore", getI64Decoder()],
    ["notAfter", getOptionDecoder(getI64Decoder())],
    ["salt", getU64Decoder()],
    ["limit", getLimitDecoder()],
    ["receive", getOptionDecoder(getReceiveDecoder())],
  ]);
}

export function getTermsCodec(): Codec<TermsArgs, Terms> {
  return combineCodec(getTermsEncoder(), getTermsDecoder());
}